Extract employee records' `get_sftp_connection()` into `itou.utils.asp`
Also move SFTP related fixtures to the top level conftest to match that.
